Move tests to their own unit file
These tests verify the rendering of the Jinja templates, and don't touch the
EndpointManager routines at all.  Accordingly, move them to their own unit
file.
